vehicle/viss/0130-get-success.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Signal Server Spec Test: get tests</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-helper.js"></script>

<h3> Get test: get success case </h3>
<h3> Get Signal.Drivetrain.Transmission.Speed</h3>

<h3>
<div id="log"></div>
</h3>

<script>
var t = async_test("Get success case");

var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

vehicle.onopen = function() {
  // request a signal
  var reqId = "reqid-0130";
  var path = "Signal.Drivetrain.Transmission.Speed";

  // process messages from the server
  vehicle.onmessage = function(event) {
    var msg = JSON.parse(event.data);
    t.step_func(function() {

      if (isGetSuccessResponse(reqId, msg)) {
        assert_true(true, "Get response received. value = " + msg.value);
        helper_terminate_success("Get response received. value = " + msg.value);
      } else {
        helper_terminate_failure("Get method failed");
      }

    })();
  }

  // in case server doesn't return response
  t.step_timeout(function() {
    helper_terminate_failure("Fishish by timeout");
  }, TIME_OUT_TIME);

  addLogMessage("Send get request: path="+path);
  vehicle.send('{"action":"get","path":"'+ path +'" ,"requestId":"'+reqId+'"}');
}
</script>
